+++ frontend/src/app/main/ui/settings/password.cljs
@@ -22,6 +22,9 @@
     :old-password-not-match
     (swap! form assoc-in [:errors :password-old]
            {:message (tr "errors.wrong-old-password")})
+    :email-as-password
+    (swap! form assoc-in [:errors :password-1]
+       {:message (tr "errors.email-as-password")})
 
     (let [msg (tr "generic.error")]
       (st/emit! (dm/error msg)))))